requirements_txt (3).txt
# Core Data Science
pandas==2.0.3
numpy==1.24.3

# NLP Libraries
nltk==3.8.1
spacy==3.6.0
gensim==4.3.1

# Statistical Analysis
statsmodels==0.14.0

# Deep Learning
torch==2.0.1
transformers==4.31.0

# Machine Learning
scikit-learn==1.3.0

# Distributed Computing
pyspark==3.4.1

# API Framework
fastapi==0.101.0
uvicorn==0.23.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Configuration
pyyaml==6.0.1

# Utilities
tqdm==4.66.1
requests==2.31.0

# Additional NLP
sentencepiece==0.1.99
sacremoses==0.0.53
